% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_census_data.R
\name{all_census_data}
\alias{all_census_data}
\alias{read_acs_data}
\title{Loads all ACS data into a single data frame.}
\source{
Census API

tract/bg area: https://mcdc.missouri.edu/applications/geocorr2022.html
}
\usage{
all_census_data(
  year = current_year,
  geo = "tract",
  write = FALSE,
  read = !write
)

read_acs_data(year = current_year, geo = "tract")
}
\arguments{
\item{year}{specify map year}

\item{geo}{tract or block group}

\item{write}{logical}

\item{read}{logical}
}
\value{
a data frame
}
\description{
Loads all ACS data into a single data frame.
}
\details{
To use the Census APIs, sign up for an API key. Add Census API key to .Renviron profile and call it CENSUS_KEY. censusapi will use it by
  default. Within R, run: \cr

  Add key to .Renviron \cr
  `Sys.setenv(CENSUS_KEY= 'YOURKEYHERE')`\cr
  Reload .Renviron \cr
  `readRenviron("~/.Renviron")` \cr
  Check to see that the expected key is output in your R console \cr
  `Sys.getenv("CENSUS_KEY")` \cr

  `tidycensus::census_api_key('YOURKEYHERE', overwrite = TRUE, install = TRUE)` \cr
}
