% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_create_region_designation.R
\name{create_regions}
\alias{create_regions}
\alias{rural_overlay}
\title{Create a file with regions, adjusted for rural areas.}
\source{
shape_CA_tract created using tigris package: /data-raw/R/generate_census_shapes.R

list of TCAC rural counties: https://www.treasurer.ca.gov/ctcac/Cover-memo.pdf
}
\usage{
create_regions(year = current_year, write = FALSE, read = !write)

rural_overlay(
  block_points,
  rural_area,
  create_overlay = NULL,
  year = current_year,
  collapse = TRUE
)
}
\arguments{
\item{block_points}{Block centroids generated by `read_block_centers`.}

\item{create_overlay}{allows for running only the sp::over function
separately to reduce#'  debugging time. To use, first assign
location_overlay(create_overlay = T) to a variable,
then use location_overlay(create_overlay = variable) to run the rest of the function.}

\item{collapse}{Aggregates blocks into tract and assigns tracts to rural or urban
designation. Setting this to FALSE allows the user to view the block-level designation.}

\item{rural_overlay}{A rural shapefile.}
}
\value{
a data frame
}
\description{
`create_regions` evaluates which regions that each county belongs to, then uses `rural_overlay` to pinpoint
rural tracts.
}
\details{
`rural_overlay` merges block centers with the rural_shapefile,
and classifies the population of any block with its centroid inside the
rural shapefile as rural. It then collapses to tract level, and any tract
with over 50 percent population rural is classified as
"Rural Areas."

`rural_overlay` is executed within the `create_regions` function, and is
separated for testing convenience only. For creating data, only
`create_regions` is necessary to run.
}
