% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dict_2024.R
\docType{data}
\name{data_dict_2024}
\alias{data_dict_2024}
\title{AFFH Mapping Tool data dictionary.  Includes all variables included in the 2024 mapping interface.}
\usage{
?data_dict_2024
}
\description{

}
\details{

}
\keyword{datasets}
