% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dict_2025.R
\docType{data}
\name{data_dict_2025}
\alias{data_dict_2025}
\title{AFFH Mapping Tool data dictionary.  Includes all variables included in the 2025 mapping interface.}
\usage{
?data_dict_2025
}
\description{

}
\details{

}
\keyword{datasets}
