% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_final_public.R
\name{final_opp_public}
\alias{final_opp_public}
\title{Generates public summary files}
\usage{
final_opp_public(year = current_year, write = FALSE, change = FALSE)
}
\arguments{
\item{year}{designates the map year's filepaths}

\item{write}{whether to write new opportunity summary files}

\item{change}{whether to write new change summary files}
}
\description{
Loads final opportunity and neighborhood change data and creates excel workbook summary files, shapefiles, and data dictionaries
}
\examples{
final_opp_public(year = 2024, write = TRUE, change = TRUE)

}
