% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lihtc_distribution.R
\name{geo_lihtc}
\alias{geo_lihtc}
\title{Plot 9% and 4% large-family, new construction LIHTC across opportunity/segregation categories pre- and post-policy adoption}
\source{
TCAC Projects: https://www.treasurer.ca.gov/ctcac/projects.asp

OBI Categorical segregation: https://berkeley.app.box.com/file/1298215376451
}
\usage{
geo_lihtc(year = current_year)
}
\arguments{
\item{year}{designates the appropriate filepaths}
}
\description{
Plot 9% and 4% large-family, new construction LIHTC across opportunity/segregation categories pre- and post-policy adoption
}
\details{
Cutoff year is 2019 for 9% projects and 2021 for 4% projects.  Currently, we include projects
   with application years 2015 or newer.  geo_lihtc() loads, geocodes, and performs the spatial join of lihtc
   developments to neighborhood categories.  The output of geo_lihtc() is two intermediate csv files, which are
   read by plot_lihtc(), summarized, plotted, and the figures are saved as .tiff files in the output/year directory.
}
\examples{
geo_lihtc(year = current_year) # call first
plot_lihtc(year = current_year) # call second


}
