% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lihtc_distribution.R
\name{get_devels}
\alias{get_devels}
\title{Evaluate 9% new construction LIHTC across opportunity categories pre- and post-policy adoption}
\source{
TCAC Projects: https://www.treasurer.ca.gov/ctcac/projects.asp
}
\usage{
get_devels(year = current_year)
}
\description{
Evaluate 9% new construction LIHTC across opportunity categories pre- and post-policy adoption
}
