% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_education_data.R
\name{graduation_rates}
\alias{graduation_rates}
\title{Imports graduation data and saves the relevant variables}
\source{
https://www.cde.ca.gov/ds/sd/sd/filescohort.asp
}
\usage{
graduation_rates(year = current_year, write = FALSE, read = !write)
}
\description{
Graudation rates are calculated as the enrollment-weighted average
  of the three schools closest to the tract centroid.
}
\details{
2019 version allocates schools to school districts, but 2020 code
  simply reads in lat/lon data, as it they are distributed by allocating
  the three closest schools to the block group, as with primary schoool
  data.
}
\note{
In 2020+, DASS schools are omitted. Charter schools are included.
}
