% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{helper_functions
capture the current year for default function years}
\alias{helper_functions
capture the current year for default function years}
\alias{current_year}
\title{Reader functions for zipped files}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
current_year
}
\arguments{
\item{name}{File within zip directory.}

\item{pattern}{Name of the shapefile}

\item{filename}{Name of file being written}

\item{type}{Can be 'excel', 'csv', 'tsv', or 'table.'}

\item{folder}{Name of the subfolder for writing data. Always resides in data directory unless 'year' is NULL.}

\item{explicitNA}{If not null, retitle NA factor to the provided string.}

\item{...}{Arguments to pass through to read.xlsx, read_csv, read_tsv, or read_txt from openxlsx and readr.}
}
\description{
Reader functions for zipped files
}
\details{
\describe{
 \item{`delete_temp`}{Deletes all files in the temporary directory. }
 \item{`read_zip`}{Deploys readr and openxlsx functions on zipped files}
 \item{`filepaths`}{Creates filenames for variables used within package for easy reading
   across years. Returns variables for reading when testing package functions; if used
   externally, they will be used in the global environment}
}
}
\keyword{datasets}
