% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_enviroscreen.R
\name{read_ces}
\alias{read_ces}
\title{Imports CalEnviroScreen 4.0 data and site-based measurements (point sources of pollution)}
\usage{
read_ces(year = current_year, write = FALSE, read = !write)
}
\description{
Imports CalEnviroScreen 4.0 data and site-based measurements (point sources of pollution)
}
