% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_education_data.R
\name{read_educ_pov}
\alias{read_educ_pov}
\title{Imports education data and writes the relevant variables to the intermediate directory}
\source{
School data: https://www.cde.ca.gov/ds/si/ds/pubschls.asp

Test data: https://caaspp-elpac.ets.org/caaspp/ResearchFileListSB?

FRPM: http://www.cde.ca.gov/ds/sd/sd/filessp.asp

Cohort Grad data: https://www.cde.ca.gov/ds/ad/filesacgr.asp
}
\usage{
read_educ_pov(year = current_year)
}
\arguments{
\item{year}{designates the appropriate filepaths}

\item{write}{write the intermediate file}

\item{read}{read an existing intermediate file}
}
\value{
a data frame
}
\description{
Imports education data and writes the relevant variables to the intermediate directory
}
\details{
Reading data is weighted by total enrollment of schools that
  return 4th-grade test scores. FRPM data is limited to and weighted
  by enrollment in schools that serve elementary school students. Grad
  data is weighted by high school enrollment.
}
\examples{
graduation_rates() # reads an existing intermediate file at the default year


}
