% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_neighborhood_change.R
\name{read_neighborhood_change}
\alias{read_neighborhood_change}
\title{Import Neighborhood Change Data}
\usage{
read_neighborhood_change(year = current_year)
}
\arguments{
\item{year}{designates the map year's filepaths}
}
\description{
Imports neighborhood change data generated by CHPC identifying non-rural tracts that
have experienced both long-term (since 2000) and recent change (since 2013)
racial/ethnic and economic change.
}
\examples{
neighborhood_change(year = 2024) # imports neighborhood change output and reduces to necessary variables




}
\author{
Matt Alvarez-Nissen `mnissen@chpc.net`
}
