% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_census_geography.R
\name{read_tract_centers}
\alias{read_tract_centers}
\title{Reads pop-weighted centroids of Census tracts, block groups, and blocks}
\source{
tract/bg centroids: https://www.census.gov/geographies/reference-files/time-series/geo/centers-population.html

block centroids: https://mcdc.missouri.edu/applications/geocorr2022.html
}
\usage{
read_tract_centers(as_shape = FALSE, year = current_year)
}
\arguments{
\item{as_shape}{returns shapefile}

\item{year}{designates the appropriate filepaths}
}
\value{
a data frame
}
\description{
Reads pop-weighted centroids of Census tracts, block groups, and blocks
}
\examples{
read_tract_centers(as_shape = T) # returns tract centers as shapefile based on the default year

}
