% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_education_distance.R
\name{school_distances}
\alias{school_distances}
\title{Creates tract/block group education scores}
\usage{
school_distances(
  year = current_year,
  geo = "tract",
  write = FALSE,
  read = !write,
  testing_handle = FALSE
)
}
\arguments{
\item{year}{designates the appropriate filepaths}

\item{geo}{tract or bg}

\item{write}{write the intermediate file}

\item{read}{read an existing intermediate file}
}
\value{
a data frame
}
\description{
Creates tract/block group education scores
}
\details{
Finds school distance to tract or block group centroids, and averages
  the reading and math scores, frpm, and graduation rates of the three nearest
  schools.  4th grade and FRPM are weighted by the enrollment of schools that
  serve 4th-graders, and graduation rates are weighted by cohort size.
}
\examples{
school_distances(year = 2024, geo = 'tract', write = TRUE, read = FALSE) # writes a new file to the intermediate directory

}
