% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_tribal_land.R
\name{tribal_overlap}
\alias{tribal_overlap}
\title{Create tribal land flag}
\usage{
tribal_overlap()
}
\value{
a dataframe
}
\description{
Imports tribal lands under of the control of federally-recognized tribes, computes intersection with Census tracts, and flags
any tract where at least 25 percent of the geography’s land area is within federally-recognized tribal lands.  In final_data.R,
High-Poverty & Segregated is not assessed in tracts where the tribal land flag is raised.
}
\examples{
tribal_overlap() # loads tracts with flag for tribal land


}
