% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{filepaths}
\alias{filepaths}
\title{Shortcuts to zipped files in the data-raw directory. See `?read_zip`}
\source{
\itemize{
  \item USDA_excluded: http://www.sc.egov.usda.gov/data/data_files.html, SFH/MFH Ineligibility
  \item school data: see education_enviroscreen.R
}
}
\usage{
filepaths(year = current_year)
}
\description{
Shortcuts to zipped files in the data-raw directory. See `?read_zip`
}
