% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_final_data.R
\name{final_opp}
\alias{final_opp}
\alias{final_raw}
\alias{final_prepare}
\title{Returns the final TCAC data frame output}
\usage{
final_opp(year = current_year, write = FALSE, reduced = TRUE, as_geo = FALSE)

final_raw(year = current_year, geo = "tract", write = FALSE)

final_prepare(year = current_year, geo = "tract", .data = NULL)
}
\arguments{
\item{year}{designates the map year's filepaths}

\item{write}{write the final output}

\item{as_geo}{logical to return sf object}

\item{geo}{allows for opportunity to be assessed at the tract level in urban areas and block group in rural areas}

\item{output}{`reduced` returns only variables necessary for the map.  If `full`, return all intermediate variables.}
}
\value{
a data frame
}
\description{
`final_raw` loads and combines all intermediate files with economic, education, and environmental indicators.
`final_prepare`creates the high poverty and segregated designation and flags unreliable data.
`final_opp` creates the final opportunity scores and designations.  `final_raw` and `final_prepare` are both
inputs into `final_opp`.  Only `final_opp` is necessary to run for generating new data.
}
\examples{
final_opp(year = 2024, write = TRUE) # writes the final output


}
