% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lihtc_distrib.R
\name{opp_lihtc_devels}
\alias{opp_lihtc_devels}
\title{Plot 9% and 4% large-family, new construction LIHTC across opportunity/segregation categories pre- and post-policy adoption}
\source{
CHPC Preservation Database

2025 CTCAC/HCD Opportunity Map
}
\usage{
opp_lihtc_devels()
}
\description{
Plot 9% and 4% large-family, new construction LIHTC across opportunity/segregation categories pre- and post-policy adoption
}
\details{
Cutoff year is 2019 for 9% projects and 2021 for 4% projects.  Currently, we include projects
   with application years 2015 or newer for 9%, 2017 or newer for 4%.  opp_lihtc_devels() preps the lihtc data and performs
   the spatial join of lihtc developments to neighborhood categories.  plot_lihtc_devels() reads the output of the former,
   summarizes and plots the data, and the figures are saved as .tiff files in the products/charts directory.
}
\note{
The code below was adjusted in September, 2024 for analysis to be included in the draft 2025 OM FAQ.
}
\examples{
opp_lihtc_devels() # call first
plot_lihtc_devels() # call second


}
