% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_education_distance.R
\name{school_distances}
\alias{school_distances}
\title{Imports education data and saves the relevant variables}
\usage{
school_distances(
  year = current_year,
  geo = "tract",
  write = FALSE,
  read = !write
)
}
\arguments{
\item{year}{designates the appropriate filepaths}

\item{geo}{tract or bg}

\item{write}{will write a new file to the intermediate directory}

\item{read}{rill read an existing file from the intermediate directory}
}
\value{
a data frame
}
\description{
Imports education data and saves the relevant variables
}
\details{
Finds school distance to tract or block group centroids (within-county), and averages
  the reading and math scores, frpm, and graduation rates of three nearest
  schools. 4th grade and FRPM are weighted by the enrollment of schools that
  serve 4th-graders, and graduation rates are weighted by cohort size.
}
\examples{
school_distances(year = 2024, geo = 'tract', write = TRUE, read = FALSE) # writes a new intermediate file to the intermediate directory

}
\seealso{
education_data.R for data imported into the distance calculation
}
